Make gitbox-create more idempotent

+++ gitbox/__init__.py
@@ -7,16 +7,29 @@ import json
 import shutil
 from collections import defaultdict
 
-from .hook import CONF_FILE
-from .unbox import prompt, promptyn
+from .unbox import prompt, promptyn, append, load_conf, CONF_FILE
+
+
+def remove_duplicate_commands(commands):
+    """ Remove duplicate commands from a list """
+    seen = set()
+    index = 0
+    while index < len(commands):
+        command = tuple(commands[index])
+        if command in seen:
+            del commands[index]
+            continue
+        seen.add(command)
+        index += 1
 
 
 def copy_static(name, dest):
     """ Copy one of the static files to a destination """
-    src = os.path.join(os.path.dirname(__file__), os.pardir, 'pylint')
-    # Check the paths so there's no crash when gitbox boxes itself
-    if os.path.abspath(src) != os.path.abspath(dest):
-        shutil.copyfile(os.path.join(src, name), os.path.join(dest, name))
+    destfile = os.path.join(dest, name)
+    if not os.path.exists(destfile):
+        srcfile = os.path.join(os.path.dirname(__file__), os.pardir, 'pylint',
+                               name)
+        shutil.copyfile(srcfile, destfile)
 
 
 def configure(repo):
@@ -26,7 +39,9 @@ def configure(repo):
                           "hooks will require gitbox to be "
                           "installed\nStandalone mode?", True)
 
-    conf = {'all': []}
+    conf = load_conf()
+    conf.setdefault('all', [])
+    conf.setdefault('modified', {})
 
     env = prompt("Path of virtualenv (relative to repository root)? "
                  "[venv]", 'venv')
@@ -37,6 +52,7 @@ def configure(repo):
     conf['autoenv'] = promptyn("Use autoenv?", True)
 
     modified = defaultdict(list)
+    modified.update(conf['modified'])
     requirements = []
     install_pylintrc = False
     install_pep8 = False
@@ -58,12 +74,13 @@ def configure(repo):
     if promptyn("Add autoPEP8 command?", True):
         requirements.append('autopep8')
         autopep8 = os.path.join(repo, 'run_autopep8.sh')
-        with open(autopep8, 'w') as outfile:
-            outfile.write("#!/bin/bash -e\n")
-            outfile.write("find %s -name '*.py' | xargs autopep8 -i "
-                          "--ignore=E501,E24" % package)
-        st = os.stat(autopep8)
-        os.chmod(autopep8, st.st_mode | stat.S_IEXEC)
+        if not os.path.exists(autopep8):
+            with open(autopep8, 'w') as outfile:
+                outfile.write("#!/bin/bash -e\n")
+                outfile.write("find %s -name '*.py' | xargs autopep8 -i "
+                              "--ignore=E501,E24" % package)
+            st = os.stat(autopep8)
+            os.chmod(autopep8, st.st_mode | stat.S_IEXEC)
 
     if promptyn("Prohibit trailing whitespace?", True):
         pre_commit.append("git diff-index --check --cached HEAD --")
@@ -77,6 +94,7 @@ def configure(repo):
     if not os.path.exists(hookdir):
         os.makedirs(hookdir)
 
+    # Construct the proper pre-commit hook command
     if standalone:
         hookfile = os.path.join(hookdir, 'hook.py')
         shutil.copyfile(os.path.join(os.path.dirname(__file__), 'hook.py'),
@@ -88,21 +106,22 @@ def configure(repo):
 
     if not os.path.isabs(env):
         hook_cmd = os.path.join('.', hook_cmd)
-
     pre_commit.append(hook_cmd)
 
-    precommit = os.path.join(hookdir, 'pre-commit')
-    with open(precommit, 'w') as outfile:
-        outfile.write("#!/bin/bash -e\n")
-        outfile.write('\n'.join(pre_commit))
-    st = os.stat(precommit)
-    os.chmod(precommit, st.st_mode | stat.S_IEXEC)
+    # Write the pre-commit file
+    precommit_file = os.path.join(hookdir, 'pre-commit')
+    if not os.path.exists(precommit_file):
+        with open(precommit_file, 'w') as outfile:
+            outfile.write("#!/bin/bash -e\n")
+            outfile.write('\n'.join(pre_commit))
+    else:
+        append(pre_commit, precommit_file)
+    st = os.stat(precommit_file)
+    os.chmod(precommit_file, st.st_mode | stat.S_IEXEC)
 
     # Write the required packages to a file
     if requirements:
-        require_file = os.path.join(repo, 'requirements_dev.txt')
-        with open(require_file, 'w') as outfile:
-            outfile.write('\n'.join(requirements))
+        append(requirements, os.path.join(repo, 'requirements_dev.txt'))
 
     # Create the pylint & pep8 config files
     if install_pylintrc or install_pep8:
@@ -115,18 +134,24 @@ def configure(repo):
             copy_static('pep8.ini', pylintdir)
 
     if conf.get('autoenv'):
-        with open(os.path.join(repo, '.env'), 'w') as outfile:
-            if os.path.isabs(env):
-                outfile.write(r'source ' + os.path.join(env, 'bin',
-                                                        'activate'))
-            else:
-                outfile.write(r"_envdir=$(dirname "
-                              r"${_files[_file-__array_offset]})")
-                outfile.write('\n')
-                outfile.write(r'source $_envdir/' + os.path.join(env, 'bin',
-                                                                 'activate'))
+        envfile = os.path.join(repo, '.env')
+        if not os.path.exists(envfile):
+            with open(envfile, 'w') as outfile:
+                if os.path.isabs(env):
+                    outfile.write(r'source ' + os.path.join(env, 'bin',
+                                                            'activate'))
+                else:
+                    outfile.write(r"_envdir=$(dirname "
+                                  r"${_files[_file-__array_offset]})")
+                    outfile.write('\n')
+                    outfile.write(r'source $_envdir/' +
+                                  os.path.join(env, 'bin', 'activate'))
 
-    conf['modified'] = dict(modified)
+    # Remove duplicates from commands
+    remove_duplicate_commands(conf['all'])
+    for key, value in modified.items():
+        remove_duplicate_commands(value)
+        conf['modified'][key] = value
     conf_file = os.path.join(repo, CONF_FILE)
     with open(conf_file, 'w') as outfile:
         json.dump(conf, outfile)
